refactor(store): convert to package + extract _common helpers (#280)
First step of the VstashStore decomposition (#280). Pure structural move,
no behavior change: all 1227 existing tests pass, plus 17 new surface guards.

- Convert vstash/store.py -> vstash/store/__init__.py (package). Relative
  imports adjusted from `.` to `..` for the deeper location; the public
  `vstash.store` import path is unchanged.
- Extract module-level constants, the pure helpers (_serialize/_deserialize/
  _cosine_sim/_normalize_tags/relevance_tier/_dot_product) and _PipelineTracer
  into vstash/store/_common.py — the leaf module imports nothing from the
  package, so no import cycles. Bodies moved byte-for-byte.
- Add __all__ to __init__.py as the explicit re-export hub so the symbols
  other modules and tests import from `vstash.store` keep resolving.
- Add tests/test_store_public_surface.py locking that re-export surface so a
  later extraction PR fails here rather than at a distant import.
